+++ lib/src/theme/widgets/flutter_deck_footer_theme.dart
@@ -24,8 +24,6 @@ class FlutterDeckFooterThemeData {
     this.slideNumberTextStyle,
     this.socialHandleColor,
     this.socialHandleTextStyle,
-    this.textColor,
-    this.textTextStyle,
   });
 
   /// Color of the slide number.
@@ -40,12 +38,6 @@ class FlutterDeckFooterThemeData {
   /// Text style of the social handle.
   final TextStyle? socialHandleTextStyle;
 
-  /// Color of the footer text.
-  final Color? textColor;
-
-  /// Text style of the footer text.
-  final TextStyle? textTextStyle;
-
   /// Creates a copy of this object with the given fields replaced with the new
   /// values.
   FlutterDeckFooterThemeData copyWith({
@@ -53,8 +45,6 @@ class FlutterDeckFooterThemeData {
     TextStyle? slideNumberTextStyle,
     Color? socialHandleColor,
     TextStyle? socialHandleTextStyle,
-    Color? textColor,
-    TextStyle? textTextStyle,
   }) {
     return FlutterDeckFooterThemeData(
       slideNumberColor: slideNumberColor ?? this.slideNumberColor,
@@ -62,8 +52,6 @@ class FlutterDeckFooterThemeData {
       socialHandleColor: socialHandleColor ?? this.socialHandleColor,
       socialHandleTextStyle:
           socialHandleTextStyle ?? this.socialHandleTextStyle,
-      textColor: textColor ?? this.textColor,
-      textTextStyle: textTextStyle ?? this.textTextStyle,
     );
   }
 
@@ -80,9 +68,6 @@ class FlutterDeckFooterThemeData {
       socialHandleTextStyle:
           socialHandleTextStyle?.merge(other.socialHandleTextStyle) ??
               other.socialHandleTextStyle,
-      textColor: other.textColor,
-      textTextStyle:
-          textTextStyle?.merge(other.textTextStyle) ?? other.textTextStyle,
     );
   }
 }