+++ lib/src/widgets/flutter_deck_footer.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_deck/src/flutter_deck.dart';
 import 'package:flutter_deck/src/flutter_deck_configuration.dart';
 import 'package:flutter_deck/src/flutter_deck_layout.dart';
+import 'package:flutter_deck/src/theme/flutter_deck_theme.dart';
 import 'package:flutter_deck/src/theme/widgets/flutter_deck_footer_theme.dart';
 
 /// A widget that renders a footer for a slide. The footer can contain the slide
@@ -21,8 +22,6 @@ import 'package:flutter_deck/src/theme/widgets/flutter_deck_footer_theme.dart';
 ///     socialHandleColor: Colors.white,
 ///     socialHandleTextStyle:
 ///         FlutterDeckTheme.of(context).textTheme.bodyMedium,
-///     textColor: Colors.white,
-///     textTextStyle: FlutterDeckTheme.of(context).textTheme.bodySmall,
 ///   ),
 ///   child: FlutterDeckFooter.fromConfiguration(
 ///     configuration: const FlutterDeckFooterConfiguration(),
@@ -36,13 +35,14 @@ class FlutterDeckFooter extends StatelessWidget {
   /// default, it is false.
   ///
   /// If [showSocialHandle] is true, the speaker's social handle will be
-  /// rendered. By default, it is false.
+  /// rendered. By default, it is false. If [widget] is also provided, [widget]
+  /// will be rendered instead.
   ///
-  /// if [text] is not null, it will be rendered. By default, it is null.
+  /// if [widget] is not null, it will be rendered. By default, it is null.
   const FlutterDeckFooter({
     this.showSlideNumber = false,
     this.showSocialHandle = false,
-    this.text,
+    this.widget,
     super.key,
   });
 
@@ -54,7 +54,7 @@ class FlutterDeckFooter extends StatelessWidget {
     super.key,
   })  : showSlideNumber = configuration.showSlideNumbers,
         showSocialHandle = configuration.showSocialHandle,
-        text = configuration.text;
+        widget = configuration.widget;
 
   /// Whether to show the slide number in the footer.
   final bool showSlideNumber;
@@ -63,7 +63,7 @@ class FlutterDeckFooter extends StatelessWidget {
   final bool showSocialHandle;
 
   /// A text to show in the footer.
-  final String? text;
+  final Widget? widget;
 
   @override
   Widget build(BuildContext context) {
@@ -75,7 +75,7 @@ class FlutterDeckFooter extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         crossAxisAlignment: CrossAxisAlignment.end,
         children: [
-          if (showSocialHandle)
+          if (showSocialHandle && widget == null)
             Text(
               context.flutterDeck.speakerInfo?.socialHandle ?? '',
               style: theme.socialHandleTextStyle?.copyWith(
@@ -83,12 +83,10 @@ class FlutterDeckFooter extends StatelessWidget {
                 fontWeight: FontWeight.bold,
               ),
             ),
-          if (text != null)
-            Text(
-              text!,
-              style: theme.textTextStyle?.copyWith(
-                color: theme.textColor,
-              ),
+          if (widget != null)
+            DefaultTextStyle(
+              style: FlutterDeckTheme.of(context).textTheme.bodySmall,
+              child: widget!,
             ),
           if (showSlideNumber)
             Text(